Amend tomcat7-maven-plugin configuration to include necessary dependencies
Since MUMUP-1816 was applied to angularjs-portal-frame, jstl is no
longer available in the war. Jetty doesn't mind, but Tomcat 7 does. We
don't need JSTL in the war at all, so just adding it to the war
dependencies is out of the question. This change adds the needed jars to
the tomcat7-maven-plugin runtime classpath.

http://stackoverflow.com/questions/6094329/tomcat-7-and-jstl was very
helpful, specifically the following:

"These dependencies are very important. JSTL 2.1 + Tomcat 7 + Servlet
3.0 is very broken unless you fix it by using these lines, especially
the exclusion part. What is happening is the JSTL 2.1 Jars are actually
pulling in the wrong versions of the Servlet spec--2.5. Unless you stop
that from happening, you will be in a whole world of pain. "